+++ package.json
@@ -48,6 +48,7 @@
     "jsdoc-react-proptypes": "^1.0.2",
     "jsdom": "^9.8.3",
     "jsdonk": "^0.5.0",
+    "lru-memoize": "^1.0.2",
     "nyc": "^10.0.0",
     "react": "^15.3.2",
     "react-addons-test-utils": "^15.0.0",
@@ -55,6 +56,7 @@
     "sinon": "^1.17.6"
   },
   "peerDependencies": {
+    "lru-memoize": "^1.0.2",
     "react": "^15.x.x"
   },
   "dependencies": {
@@ -64,7 +66,6 @@
     "element-resize-detector": "^1.1.12",
     "immutable": "^3.8.1",
     "is-plain-object": "^2.0.3",
-    "lru-memoize": "^1.0.2",
     "moment": "^2.18.1",
     "numeral": "^2.0.6",
     "prop-types": "^15.5.8",
