+++ src/deprecated/pipe/SplitIndexPipe-test.js
