+++ src/deprecated/pipe/SplitKeyPipe-test.js
