+++ src/deprecated/pipe/SpreadPipe-test.js
