--- src/index.js
@@ -1,31 +1,0 @@
-export {default as Button} from './component/button/Button';
-export {default as Table} from './component/table/Table';
-export {default as Label} from './component/field/Label';
-export {default as ShowHide} from './component/showHide/ShowHide';
-export {ShowHideStateful} from './component/showHide/ShowHide';
-
-export {default as Input} from './input/input/Input';
-export {default as Textarea} from './input/textarea/Textarea';
-export {default as Select} from './input/select/Select';
-export {default as Toggle} from './input/toggle/Toggle';
-export {default as ToggleSet} from './input/toggleSet/ToggleSet';
-
-export {default as SpruceClassName} from './util/SpruceClassName';
-export {default as SpruceComponent} from './util/SpruceComponent';
-export {default as CollectionUtils} from './util/CollectionUtils';
-export {default as Compose} from './util/Compose';
-export {default as ConfigureHock} from './util/ConfigureHock';
-
-export {default as ElementQueryHock} from './hock/ElementQueryHock';
-export {default as PropChangeHock} from './hock/PropChangeHock';
-export {default as QueryStringHock} from './hock/QueryStringHock';
-export {default as StateHock} from './hock/StateHock';
-
-
-// pipes
-export {default as SpreadPipe} from './pipe/SpreadPipe';
-export {default as SplitKeyPipe} from './pipe/SplitKeyPipe';
-export {default as SplitIndexPipe} from './pipe/SplitIndexPipe';
-export {default as DownPipe} from './pipe/DownPipe';
-export {default as UpPipe} from './pipe/UpPipe';
-export {default as DebouncePipe} from './pipe/DebouncePipe';
